xinziweb_v1_0/requirements.txt
Django=3.2.23	
PyJWT=1.7.1	
PyMySQL=1.1.0	
asgiref=3.7.2	
djangorestframework=3.14.0	
djangorestframework-jwt=1.11.0	
djangorestframework-simplejwt=5.3.0	
pip=23.3.1
pytz=2023.3.post1	
setuptools=68.0.0	
sqlparse=0.4.4	
typing-extensions=4.7.1	
wheel=0.41.2	
